requirements.txt
asgiref
Django
django-ckeditor
django-js-asset
django-mptt
Pillow
sqlparse
